+++ src/main/php/CatchedException.php
@@ -65,7 +65,7 @@ class CatchedException
      *
      * @api
      */
-    public function with(Predicate|callable $predicate, string $description = null): self
+    public function with(Predicate|callable $predicate, ?string $description = null): self
     {
         assertThat($this->actualException, $predicate, $description);
         return $this;
@@ -76,7 +76,7 @@ class CatchedException
      *
      * @api
      */
-    public function after(mixed $value, Predicate|callable $predicate, string $description = null): self
+    public function after(mixed $value, Predicate|callable $predicate, ?string $description = null): self
     {
         assertThat($value, $predicate, $description);
         return $this;